backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
mysql-connector-python

groq
requests==2.31.0
pillow==10.1.0
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-magic==0.4.27
opencv-python
email-validator==2.1.0
pydantic[email]==2.5.0
mediapipe==0.10.8
ultralytics 
websockets